requirements.txt
# Sports Analytics Agent - Python Dependencies
# Install: pip install -r requirements.txt
# Python 3.10+

# UI
streamlit>=1.35.0

# LLM / Agent Framework
langchain>=0.2.0
langchain-groq>=0.1.6
langchain-community>=0.2.0
langchain-chroma>=0.1.2
langchain-neo4j>=0.1.0
langchain-text-splitters>=0.2.0

# Databases
neo4j>=5.19.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# Data
pandas>=2.0.0

# Embeddings and Vector Store
sentence-transformers>=3.0.0
chromadb>=0.5.0

# PDF Loading
pypdf>=4.0.0

# Vision (Gemini)
google-genai>=1.0.0
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
